Extract training logic into dedicated module
